requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
psycopg2-binary
python-dotenv
bcrypt
marshmallow
flask_jwt_extended 